refactor: move format and genres inputs into separate components
