+++ src/components/form/index.ts
@@ -1,0 +1,5 @@
+export { default as DateInput } from './DateInput/DateInput';
+export { default as Format } from './Format/Format';
+export { default as Genres } from './Genres/Genres';
+export { default as Select } from './Select/Select';
+export { default as Title } from './Title/Title';
