+++ src/components/index.ts
@@ -4,7 +4,6 @@ export { default as CardContainer } from './cardContainer/CardContainer';
 export { default as Modal } from './modal/Modal';
 export { default as Navbar } from './navbar/Navbar';
 export { default as Form } from './form/Form';
-export { default as Select } from './form/Select';
 export { default as Header } from './header/Header';
 export { default as SearchBar } from './searchBar/SearchBar';
 export { default as Spinner } from './spinner/Spinner';
